+++ packages/template/src/types/i18n.d.ts
@@ -11,4 +11,3 @@ declare module '@/page/*/i18n/*.json' {
   const value: Record<string, any>
   export default value
 }
-